+++ backend-service/src/main/java/com/heulwen/backendservice/service/UserService.java
@@ -4,6 +4,7 @@ import com.heulwen.backendservice.dto.UserDto;
 import com.heulwen.backendservice.form.UserCreateForm;
 import com.heulwen.backendservice.form.UserUpdateForm;
 import com.heulwen.backendservice.model.User;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
@@ -11,8 +12,8 @@ import java.util.List;
 public interface UserService extends UserDetailsService {
     UserDto createUser(UserCreateForm form, MultipartFile avatar);
     UserDto createAdmin(UserCreateForm form, MultipartFile avatar);
-    UserDto updateUser(Integer userId, UserUpdateForm form, MultipartFile avatar);
-    void deleteAdmin(int id);
+    UserDto updateUser(Long userId, UserUpdateForm form, MultipartFile avatar);
+    void deleteAdmin(Long id);
     List<UserDto> getUsersAdmin(String role);
     List<UserDto> getUsers();
     User getUserByUsername(String username);
